+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
   * Added `--protocol-path` parameter for limiting protocol paths to be executed (#23)
   * Set default price to 1 ETH (#24)
+  * Output Format: Add scaling support (#25)
 
 ## v1.1.1