+++ CHANGELOG.md
@@ -5,6 +5,9 @@
   * Added `--protocol-path` parameter for limiting protocol paths to be executed (#23)
   * Set default price to 1 ETH (#24)
   * Output Format: Add scaling support (#25)
+  * Protocol: Delgado and Delgado-Reusable (#1)
+  * Documentation of Delgado updated
+  * Dependency Update: ecdsa 0.15
 
 ## v1.1.1